Added Plugins folder for Git LFS functionality directly within UE4
This plugin works on a per-project basis.  When connecting to source control, it will show Git LFS instead of the normal Git (beta).  This will ONLY affect this project in UE4.  Connecting to source control on any other project that doesn't include the contents of this Plugins folder will not show Git LFS and provide the extra functionality.